Crate octavo

Highly modular & configurable hash & crypto library written in Rust

About

Octavo is Rust implementation of cryptographic primitives and Transport Layer Security. Our goal is to provide safe, fast, full-featured and configurable cryptographic library for modern world.

Except of Rust API we want expose also C API for usage with other languages and tools. Probably there will also land thin OpenSSL compatibility cascade for use with existing code.

Legalities

Please remember that export/import and/or use of strong cryptography software, providing cryptography hooks, or even just communicating technical details about cryptography software is illegal in some parts of the world. So when you import this package to your country, re-distribute it from there or even just email technical suggestions or even source patches to the authors or other people you are strongly advised to pay close attention to any laws or regulations which apply to you. The authors of Octavo are not liable for any violations you make here. So be careful, it is your responsibility.1

Text of this paragraph is copied from OpenSSL website. ↩

Reexports

pub extern crate octavo_crypto;

pub extern crate octavo_digest;

pub extern crate octavo_kdf;

pub extern crate octavo_mac;

Modules

crypto

digest

kdf

mac