Module octavo::digest::prelude::sha2

SHA-2 family (Secure Hash Algorithm)

General info

Name 	Digest size 	Block size 	Rounds 	Structure 	Reference

SHA-224 	224 bits 	512 bits 	64 	Merkle–Damgård 	FIPS 180-4

SHA-256 	256 bits 	512 bits 	64 	Merkle–Damgård 	FIPS 180-4

SHA-384 	384 bits 	1024 bits 	80 	Merkle–Damgård 	FIPS 180-4

SHA-512 	512 bits 	1024 bits 	80 	Merkle–Damgård 	FIPS 180-4

SHA-512/224 	224 bits 	1024 bits 	80 	Merkle–Damgård 	FIPS 180-4

SHA-512/256 	256 bits 	1024 bits 	80 	Merkle–Damgård 	FIPS 180-4

Warning

SHA-384, SHA-512, SHA-512/224 and SHA-512/256 implementations are specified to manage messages with legth at most 2128 - 1 bits length. Due to simplification of implementation this library currently manage messages with length at most 264 - 1 bits which is reasonable for most usages (2 EB which is greater than maximum partition size in ext4).

Structs

Sha224
SHA-224 implementation

Sha256
SHA-256 implementation

Sha384
SHA-384 implementation

Sha512
SHA-512 implementation

Sha512224
SHA-512/224 implementation

Sha512256
SHA-512/256 implementation